+++ example/my_request/my_interface.dart
@@ -4,4 +4,5 @@ import 'package:wordpress_client/wordpress_client.dart';
 import 'my_request.dart';
 import 'my_response.dart';
 
-class MyInterface extends IInterface with ListMixin<MyResponse, MyRequest> {}
+class MyInterface extends IWordpressService
+    with ListMixin<MyResponse, MyRequest> {}