+++ lib/src/authorization/authorization_builder.dart
@@ -3,9 +3,9 @@
 import '../enums.dart';
 import '../utilities/wordpress_callback.dart';
 import '../wordpress_client_base.dart';
-import 'authorization_methods/basic_auth.dart';
-import 'authorization_methods/basic_jwt.dart';
-import 'authorization_methods/useful_jwt.dart';
+import 'types/basic_auth.dart';
+import 'types/basic_jwt.dart';
+import 'types/useful_jwt.dart';
 
 class AuthorizationBuilder {
   String _userName = '';