+++ lib/src/authorization/authorization_export.dart
@@ -1,4 +1,4 @@
 export 'authorization_builder.dart';
-export 'authorization_methods/basic_auth.dart';
-export 'authorization_methods/basic_jwt.dart';
-export 'authorization_methods/useful_jwt.dart';
+export 'types/basic_auth.dart';
+export 'types/basic_jwt.dart';
+export 'types/useful_jwt.dart';