+++ lib/src/library_exports.dart
@@ -5,4 +5,4 @@ export 'requests/request_interface.dart';
 export 'requests/wordpress_request.dart';
 export 'responses/wordpress_response.dart';
 export 'wordpress_client_base.dart'
-    show IAuthorization, IInterface, WordpressClient;
+    show IAuthorization, IWordpressService, WordpressClient;