+++ lib/src/responses/error_container.dart
@@ -1,7 +1,7 @@
 import 'package:dio/dio.dart';
 
 class ErrorContainer {
-  ErrorContainer({this.errorResponse, this.internalError});
+  const ErrorContainer({this.errorResponse, this.internalError});
 
   final dynamic errorResponse;
   final DioError? internalError;