+++ lib/src/responses/post_response.dart
@@ -1,8 +1,8 @@
 import '../enums.dart';
 import '../utilities/helpers.dart';
-import 'response_properties/author_meta.dart';
-import 'response_properties/content.dart';
-import 'response_properties/links.dart';
+import 'properties/author_meta.dart';
+import 'properties/content.dart';
+import 'properties/links.dart';
 
 class Post {
   Post({