+++ lib/src/responses/tag_response.dart
@@ -1,7 +1,8 @@
-import 'response_properties/links.dart';
+import '../interfaces/serializable_base.dart';
+import 'properties/links.dart';
 
-class Tag {
-  Tag({
+class Tag extends ISerializable<Tag> {
+  const Tag({
     this.id,
     this.count,
     this.description,
@@ -49,5 +50,54 @@ class Tag {
       'meta': meta,
       '_links': links?.toJson(),
     };
+  }
+
+  @override
+  Tag fromMap(Map<String, dynamic> map) {
+    return Tag.fromJson(map);
+  }
+
+  @override
+  Map<String, dynamic> toMap() {
+    return toJson();
+  }
+
+  @override
+  List<Object?> get props {
+    return [
+      id,
+      count,
+      description,
+      link,
+      name,
+      slug,
+      taxonomy,
+      meta,
+      links,
+    ];
+  }
+
+  Tag copyWith({
+    int? id,
+    int? count,
+    String? description,
+    String? link,
+    String? name,
+    String? slug,
+    String? taxonomy,
+    dynamic meta,
+    Links? links,
+  }) {
+    return Tag(
+      id: id ?? this.id,
+      count: count ?? this.count,
+      description: description ?? this.description,
+      link: link ?? this.link,
+      name: name ?? this.name,
+      slug: slug ?? this.slug,
+      taxonomy: taxonomy ?? this.taxonomy,
+      meta: meta ?? this.meta,
+      links: links ?? this.links,
+    );
   }
 }