+++ lib/src/responses/wordpress_response.dart
@@ -1,44 +1,69 @@
 import '../utilities/helpers.dart';
 
 class WordpressResponse<T> {
-  WordpressResponse(
+  const WordpressResponse._(
     this.data, {
     required this.responseCode,
     required this.responseHeaders,
-    this.requestDuration,
+    this.requestDuration = Duration.zero,
     this.message,
   });
 
-  WordpressResponse.success(
-    this.data, {
-    this.responseCode = 200,
-    required this.responseHeaders,
-    this.requestDuration,
-    this.message,
-  });
+  factory WordpressResponse.success({
+    required T responseData,
+    required Map<String, dynamic> headers,
+    int responseCode = 200,
+    Duration duration = Duration.zero,
+    String? message,
+  }) {
+    return WordpressResponse._(
+      responseData,
+      responseCode: responseCode,
+      responseHeaders: headers,
+      requestDuration: duration,
+      message: message,
+    );
+  }
 
-  WordpressResponse.failed(
-    this.data, {
-    this.responseCode = -1,
-    this.responseHeaders = const <String, dynamic>{},
-    this.requestDuration,
-    this.message,
-  });
+  factory WordpressResponse.failed({
+    T? responseData,
+    Map<String, dynamic> headers = const <String, dynamic>{},
+    int responseCode = -1,
+    Duration duration = Duration.zero,
+    String? message,
+  }) {
+    return WordpressResponse._(
+      responseData,
+      responseCode: responseCode,
+      responseHeaders: headers,
+      requestDuration: duration,
+      message: message,
+    );
+  }
 
-  final T data;
+  final T? data;
   final int responseCode;
   final Map<String, dynamic> responseHeaders;
-  final Duration? requestDuration;
+  final Duration requestDuration;
   final String? message;
 
   bool get isSuccess => isInRange(responseCode, 200, 299);
 
-  int get totalPagesCount => responseHeaders.isNotEmpty
-      ? int.tryParse((responseHeaders['x-wp-totalpages'] as String?) ?? '0') ??
-          0
-      : 0;
+  int get totalPagesCount {
+    if (responseHeaders.isEmpty || responseHeaders['x-wp-totalpages'] == null) {
+      return 0;
+    }
 
-  int get totalCount => responseHeaders.isNotEmpty
-      ? int.tryParse((responseHeaders['x-wp-total'] as String?) ?? '0') ?? 0
-      : 0;
+    final totalPagesHeaderValue = responseHeaders['x-wp-totalpages'] as String;
+    return int.tryParse(totalPagesHeaderValue)!;
+  }
+
+  int get totalCount {
+    if (responseHeaders.isEmpty || responseHeaders['x-wp-total'] == null) {
+      return 0;
+    }
+
+    final totalHeaderValue = responseHeaders['x-wp-total'] as String;
+    return int.tryParse(totalHeaderValue)!;
+  }
 }