+++ lib/src/services/search.dart
@@ -3,5 +3,5 @@ import '../../responses.dart';
 import '../library_exports.dart';
 import '../requests/list/list_search.dart';
 
-class SearchInterface extends IInterface
+class SearchService extends IWordpressService
     with ListMixin<Search, ListSearchRequest> {}